Refactored QueryForm.js
