+++ src/QueryForm.js
@@ -1,41 +1,11 @@
 import React, {useState} from 'react';
-import {makeStyles} from "@material-ui/core/styles";
-import {useLocation, withRouter} from "react-router-dom";
-import Accordion from "@material-ui/core/Accordion";
-import AccordionSummary from "@material-ui/core/AccordionSummary";
-import AccordionDetails from "@material-ui/core/AccordionDetails";
-import Typography from "@material-ui/core/Typography";
-import ExpandMoreIcon from '@material-ui/icons/ExpandMore';
-import FilterListIcon from '@material-ui/icons/FilterList';
+import {withRouter} from "react-router-dom";
 import Box from "@material-ui/core/Box";
-import Chip from "@material-ui/core/Chip";
 import SearchBar from "./SearchBar";
-
-const useStyles = makeStyles((theme) => ({
-    divider: {
-        height: 28,
-        margin: 4,
-    },
-    filters: {
-        width: "600px",
-        marginTop: "15px"
-    },
-    filtersSummary: {
-        color: "#535353"
-    },
-    filtersIcon: {
-        marginRight: "10px"
-    }
-}));
-
-//explicit is better than implicit
-const daysInWeek = [
-    "pondělí",
-    "úterý",
-    "středa",
-    "čtvrtek",
-    "pátek"
-];
+import DayPicker from "./DayPicker";
+import {useQuery} from "./hooks";
+import Filters from "./Filters";
+import {daysInWeek} from "./constants";
 
 const daysToInfo = (days) => {
     if (days.length === 5) return "kterýkoliv den"
@@ -45,10 +15,6 @@ const daysToInfo = (days) => {
     return preposition +" " + days;
 }
 
-const useQuery = () => {
-    return new URLSearchParams(useLocation().search);
-};
-
 const parseDays = (daysString) => {
     if (!daysString) return [];
     return [...daysString].map(letter => parseInt(letter));
@@ -56,7 +22,6 @@ const parseDays = (daysString) => {
 
 const QueryForm = ({history}) => {
     let queryRoute = useQuery();
-    const classes = useStyles();
     const [query, setQuery] = useState(queryRoute.get("includes") ? queryRoute.get("includes") : "");
     const [days, setDays] = useState(parseDays(queryRoute.get("days")));
 
@@ -72,44 +37,13 @@ const QueryForm = ({history}) => {
         if (params.charAt(0) === '&') params = params.substr(1);
         return history.push('/search?' + params);
     };
-    const handleDayChange = (day) => (e) => {
-        e.preventDefault();
-        if (days.includes(day)) {
-            setDays(days.filter((_day => day !== _day)));
-        } else {
-            setDays([...days, day]);
-        }
-    }
+
     return (
         <Box component="form" onSubmit={handleSubmit}>
-            <SearchBar value={query} onChange={(value) => {setQuery(value)}}/>
-            <Accordion
-                className={classes.filters}
-            >
-                <AccordionSummary
-                    expandIcon={<ExpandMoreIcon/>}
-                    aria-controls="filtry"
-                    className={classes.filtersSummary}
-                >
-                    <FilterListIcon className={classes.filtersIcon}/>
-                    <Typography>{
-                        days.length > 0 ? "Předměty " + daysToInfo(days) : "Filtry"
-                    }</Typography>
-                </AccordionSummary>
-                <AccordionDetails>
-                    <div>
-                        {daysInWeek.map((day, index) =>
-                            <Chip
-                                key={index}
-                                label={day}
-                                onClick={handleDayChange(index)}
-                                variant={days.includes(index) ? 'default' : "outlined"}
-                                color={days.includes(index) ? "primary" : 'default'}
-                            />
-                        )}
-                    </div>
-                </AccordionDetails>
-            </Accordion>
+            <SearchBar value={query} onChange={setQuery}/>
+            <Filters info={days.length > 0 ? "Předměty " + daysToInfo(days) : "Filtry"}>
+                <DayPicker value={days} onChange={setDays}/>
+            </Filters>
         </Box>
     );
 };
