+++ src/constants.js
@@ -1,0 +1,9 @@
+
+/** Day string s in czech */
+export const daysInWeek = [
+    "pondělí",
+    "úterý",
+    "středa",
+    "čtvrtek",
+    "pátek"
+];
